FilterKinect

Documentation for FilterKinect.

FilterKinect.filter!

FilterKinect.get_header_names

FilterKinect.getoption

FilterKinect.getrow

FilterKinect.load_markerdata

FilterKinect.save_markerdata

FilterKinect.setoption

FilterKinect.validate_max

FilterKinect.validate_min

FilterKinect.filter! — Method

filter!(data::MarkerData, header_name::String; min_frq=1, max_frq=80)

Filter column header_name of data by first calculating a frequency representation of the data using FFT and setting every frequency not between min_frq and max_frq (inclusive) to 0. Then the FFT data is reconstructed into an array of real numbers.

This filtered result is written into the corresponding column of data.filtered_df.

FilterKinect.get_header_names — Method

get_header_names(path::String)

Return vector of header names as string and the first six lines of the file as tuple.

Example:

julia> get_header_names("my_file.trc")[1]
98-element Vector{String}:
 "Frame"
 "Time"
 "PELVIS_X"
 "PELVIS_Y"
 "PELVIS_Z"
 "SPINE_NAVAL_X"
 "SPINE_NAVAL_Y"
 "SPINE_NAVAL_Z"
 ⋮
 "EAR_LEFT_Z"
 "EYE_RIGHT_X"
 "EYE_RIGHT_Y"
 "EYE_RIGHT_Z"
 "EAR_RIGHT_X"
 "EAR_RIGHT_Y"
 "EAR_RIGHT_Z"

FilterKinect.getoption — Method

getoption(filterconfig::FilterConfig, markername, option)

Return value of option option for marker markername saved in filterconfig.

FilterKinect.getrow — Method

getrow(df::DataFrame, markername)

Return the row of the given marker (as a view).

FilterKinect.load_markerdata — Method

load_marker_data(path::String)

Load coordinate data from a .trc file. Return a ModelData object.

FilterKinect.save_markerdata — Method

save_markerdata(data::MarkerData, filepath::String; overwrite = false)

Save data in file at filepath.

overwrite: If false and the file already exists or it's a folder, an exception will be thrown. If true, the file gets completely overwritten instead. The default is false.

FilterKinect.setoption — Method

setoption(filterconfig::FilterConfig, markername, option, value)

Set option option for marker markername in filterconfig to given value.

FilterKinect.validate_max — Method

validate_max(value)

Validate that new maximum frequency for filtering value is legal (not smaller than minimum slider currently has and not bigger than max_possible).

FilterKinect.validate_min — Method

validate_min(value)

Validate that new minimum frequency for filtering value is legal (not bigger than maximum slider currently has and not smaller than min_possible).